+++ SU2_CFD/src/driver_structure.cpp
@@ -3483,7 +3483,7 @@ void CDriver::XBraidPreprocessing(){
 
 #ifdef HAVE_XBRAID
 
-    int nPoint              = geometry_container[ZONE_0][INST_0][MESH_0]->GetnPoint();
+    unsigned long nPoint              = geometry_container[ZONE_0][INST_0][MESH_0]->GetnPoint();
     int nVar                = solver_container[ZONE_0][INST_0][MESH_0][FLOW_SOL]->GetnVar();
 
     /* --- Set up the application structure for xBraid --- */
@@ -3540,8 +3540,8 @@ void CDriver::XBraidPreprocessing(){
 
     /* --- Grab the initial condition from SU2 and store pointer in app --- */
     TwoStepSolution *initial_condition  = new TwoStepSolution(app->BDF2, nPoint, nVar);
-    for (int iPoint = 0; iPoint < nPoint; iPoint++){
-      for (int iVar = 0; iVar < nVar; iVar++){
+    for (unsigned long iPoint = 0; iPoint < nPoint; iPoint++){
+      for (unsigned short iVar = 0; iVar < nVar; iVar++){
         initial_condition->time_n[iPoint][iVar] = SU2_TYPE::GetValue(solver_container[ZONE_0][INST_0][MESH_0][FLOW_SOL]->node[iPoint]->GetSolution(iVar));
         if(app->BDF2) initial_condition->time_n1[iPoint][iVar] = SU2_TYPE::GetValue(solver_container[ZONE_0][INST_0][MESH_0][FLOW_SOL]->node[iPoint]->GetSolution_time_n1()[iVar]);
       }
